Feat: udistrital/sga_documentacion#25 revert migracion
revert logica ultimos controladores
